updated on few pages

+++ backend/controllers/eduContentController.js
@@ -5,6 +5,7 @@ const ApiFeatures = require("../utils/apifeatures");
 const cloudinary = require("cloudinary");
 
 exports.createEduContent = catchAsyncErrors(async (req, res, next) => {
+    req.body.postedBy = req.user.id;
     const eduContent = await EduContent.create(req.body);
     return res.status(201).json({
         success: true,
@@ -21,15 +22,18 @@ exports.getAllEduContent = catchAsyncErrors(async (req, res, next) => {
         });
     }
     return res.status(500).json({
-        success:false,
-        message:"error"
+        success: false,
+        message: "error"
     })
 });
 
 exports.updateEduContent = catchAsyncErrors(async (req, res, next) => {
     let eduContent = await EduContent.findById(req.params.id);
     if (!eduContent) {
-        return next(new ErrorHandler("EduContent not found",404));
+        return next(new ErrorHandler("EduContent not found", 404));
+    }
+    if (eduContent.postedBy !== req.user.id) {
+        return next(new ErrorHandler("You are not authorized to update this eduContent", 403));
     }
     eduContent = await EduContent.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
@@ -45,9 +49,11 @@ exports.updateEduContent = catchAsyncErrors(async (req, res, next) => {
 exports.deleteEduContent = catchAsyncErrors(async (req, res, next) => {
     let eduContent = await EduContent.findById(req.params.id);
     if (!eduContent) {
-        return next(new ErrorHandler("EduContent not found",404));
+        return next(new ErrorHandler("EduContent not found", 404));
     }
-
+    if (eduContent.postedBy !== req.user.id) {
+        return next(new ErrorHandler("You are not authorized to update this eduContent", 403));
+    }
     await eduContent.deleteOne(eduContent);
 
     return res.status(200).json({
@@ -58,7 +64,7 @@ exports.deleteEduContent = catchAsyncErrors(async (req, res, next) => {
 exports.getEduContentDetails = catchAsyncErrors(async (req, res, next) => {
     let eduContent = await EduContent.findById(req.params.id);
     if (!eduContent) {
-        return next(new ErrorHandler("EduContent not found",404));
+        return next(new ErrorHandler("EduContent not found", 404));
     }
     return res.status(200).json({
         success: true,
